最終合併 (#2)
* 更新 重新製作版面設計(偏向主題設計化)
新增 Feed css
修改 AnimalController 過濾
修改 MemberFactory 快捷

* 更新 重新製作版面設計(偏向主題設計化)
修改 Feed css
更新 Button css
更新 Form css
更新 Card css

* 更新 重新製作版面設計(偏向主題設計化)
更新 Placeholder css
更新 Card css

* 清除無用程式碼

* 整合移植環境
新增品牌網站初版(/branding)
分離頁面的所有資源加載檔案
更新部分 footer default layouts 樣板基底

* 移除無用資料

+++ app/Http/Controllers/AnimalController.php
@@ -2,8 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Lib\I18N\I18N;
+use App\Lib\Utils\EValidatorType;
+use App\Lib\Utils\ValidatorBuilder;
 use App\Models\Animal;
+use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\MessageBag;
 use Symfony\Component\HttpFoundation\Response as ResponseAlias;
 
 class AnimalController extends Controller
@@ -26,12 +32,23 @@ class AnimalController extends Controller
 
     /**
      * Store a newly created resource in storage.
+     * @throws Exception
      */
     public function store(Request $request)
     {
         //
-        $var = Animal::create($request->all());
-        return response()->json(['message' => "Animal Added"], ResponseAlias::HTTP_CREATED);
+        $init = self::baseControllerInit($request);
+        $i18N = $init['i18N'];
+        if (!($i18N instanceof I18N))  throw new Exception('$i18N Not instanceof I18N');
+        $vb = new ValidatorBuilder($i18N, EValidatorType::ANIMALCREATE);
+        $v = $vb->validate($request->all());
+
+        if($v instanceof MessageBag){
+            return response()->json(['message' => "Failed to create animal", 'errors' => [...$v->all()]], ResponseAlias::HTTP_BAD_REQUEST);
+        }else{
+            Animal::create($request->all());
+            return response()->json(['message' => "Animal Added"], ResponseAlias::HTTP_CREATED);
+        }
     }
 
     /**
@@ -65,14 +82,20 @@ class AnimalController extends Controller
     {
         //
         // 手动查找 Animal 模型实例
-        $animal = Animal::find($id);
 
-        // 检查是否找到实例
-        if (!$animal) {
-            // 如果没有找到，返回自定义错误消息
-            return response()->json(['message' => 'Animal not found with ID ' . $id], ResponseAlias::HTTP_NOT_FOUND);
+        $v = Validator::make(['id'=>$id], ["id"=>["required","integer"]]);
+        $validate = $v->validate();
+        if (!$v->errors()->any()) {
+            $animal = Animal::find($validate['id']);
+
+            // 检查是否找到实例
+            if (!$animal) {
+                // 如果没有找到，返回自定义错误消息
+                return response()->json(['message' => 'Animal not found with ID ' . $id], ResponseAlias::HTTP_NOT_FOUND);
+            }
+            $animal->delete();
+            return response()->json(['message' => "Animal Deleted"], ResponseAlias::HTTP_OK);
         }
-        $animal->delete();
-        return response()->json(['message' => "Animal Deleted"], ResponseAlias::HTTP_OK);
+        return response()->json(['message' => "id is invalid"], ResponseAlias::HTTP_OK);
     }
 }
